Warning fixes

+++ trunk/win32/libcdio-0.81/lib/driver/image/cdrdao.c
@@ -1028,7 +1028,7 @@ _read_mode1_sectors_cdrdao (void *user_data, void *data, lsn_t lsn,
   int retval;
   unsigned int blocksize = b_form2 ? M2RAW_SECTOR_SIZE : CDIO_CD_FRAMESIZE;
 
-  for (i = 0; i < nblocks; i++) {
+  for (i = 0; i < (int)nblocks; i++) {
     if ( (retval = _read_mode1_sector_cdrdao (env, 
 					    ((char *)data) + (blocksize * i),
 					    lsn + i, b_form2)) )
@@ -1092,7 +1092,7 @@ _read_mode2_sectors_cdrdao (void *user_data, void *data, lsn_t lsn,
   int i;
   int retval;
 
-  for (i = 0; i < nblocks; i++) {
+  for (i = 0; i < (int)nblocks; i++) {
     if ( (retval = _read_mode2_sector_cdrdao (env, 
 					    ((char *)data) + (CDIO_CD_FRAMESIZE * i),
 					    lsn + i, b_form2)) )
